requirements.txt
pyyaml
tabulate
requests
beautifulsoup4
